+++ apps/docs/contentlayer.config.ts
@@ -245,7 +245,7 @@ export const Components = defineDocumentType(() => ({
 
 export default makeSource({
     contentDirPath: "./content",
-    contentDirExclude: ["ai-templates"],
+    contentDirExclude: ["ai"],
     documentTypes: [Page, Tokens, Components, Icons, Guides, GettingStarted, StyledSystem],
     mdx: {
         remarkPlugins: [],