+++ packages/components/src/box/docs/dos-and-donts/box-as-inline-dont.tsx
@@ -1,0 +1,14 @@
+import { Box } from "@hopper-ui/components";
+
+export function Example() {
+    return (
+        <>
+            <Box display="flex">
+                This should just be a <code>&lt;Inline&gt;</code>
+            </Box>
+            <Box display="flex" flexDirection="column">
+                This should just be a <code>&lt;Stack&gt;</code>
+            </Box>
+        </>
+    );
+}